remove transition rules to addon
